+++ src/sim/sim_elements/VSimNode.h
@@ -63,8 +63,8 @@ public:
 
     void clearNeighbours(VLayerSequence layer) ;
 
-    double getDistance(const VSimNode* node) const ;
-    double getDistance(const QVector3D& point) const;
+    float getDistance(const VSimNode* node) const ;
+    float getDistance(const QVector3D& point) const;
     const QVector3D &getPosition() const ;
     void setPosition(const QVector3D& pos);
     void reset() override;
